+++ packages/stream_video/lib/src/stream_video.dart
@@ -199,6 +199,8 @@ class StreamVideo extends Disposable {
       ).then((_) {
         webrtcInitializationCompleter.complete();
       });
+    } else {
+      webrtcInitializationCompleter.complete();
     }
 
     final tokenProvider = switch (user.type) {