+++ packages/stream_video/CHANGELOG.md
@@ -9,6 +9,11 @@
 ### ⚠️ Deprecated
 * Deprecated `androidAudioConfiguration` in `StreamVideoOptions`. Use `audioConfigurationPolicy` instead.
 
+## 1.2.4
+
+### 🐞 Fixed
+* [Web] Fixed an issue with call being stuck during Connecting state.
+
 ## 1.2.3
 
 ### ⚡ Performance